+++ docs/release-notes.md
@@ -1,3 +1,8 @@
+### v7.0.8-v.26
+
+[Full Changelog](https://github.com/yakworks/gorm-tools/compare/v7.0.8-v.25...v7.0.8-v.26)
+- Parallel each (#346) [link](https://github.com/yakworks/gorm-tools/commit/1571e3b1960f8114d151ddc9c4cd13be0574c49d)
+
 ### v7.0.8-v.25
 
 [Full Changelog](https://github.com/yakworks/gorm-tools/compare/v7.0.8-v.24...v7.0.8-v.25)
@@ -317,7 +322,7 @@
 
 #### 6.2.2
  - 2019-05-20 - [4 commits](https://github.com/yakworks/gorm-tools/compare/v6.2.1...v6.2.2) by [Joshua B](https://github.com/basejump) - published to [![Bintray](https://img.shields.io/badge/Bintray-6.2.2-green.svg)](https://bintray.com/null/null/org.grails.plugins/6.2.2)
- - - rename asyncSupport [(#195)](https://github.com/yakworks/gorm-tools/pull/195)
+ - - rename parallelTools [(#195)](https://github.com/yakworks/gorm-tools/pull/195)
 
 #### 6.2.1
  - 2019-05-14 - [1 commit](https://github.com/yakworks/gorm-tools/compare/v6.2.0...v6.2.1) by [Joshua B](https://github.com/basejump) - published to [![Bintray](https://img.shields.io/badge/Bintray-6.2.1-green.svg)](https://bintray.com/null/null/org.grails.plugins/6.2.1)
@@ -418,11 +423,11 @@
  - quickSearchFields for Mango should have defaults and be able to be set in config [(#89)](https://github.com/yakworks/gorm-tools/issues/89)
  - @GormRepository annotation for AST [(#88)](https://github.com/yakworks/gorm-tools/issues/88)
  - Groovydocs updates [(#87)](https://github.com/yakworks/gorm-tools/pull/87)
- - AsyncSupport and GparsAsyncSupport need tests and docs [(#86)](https://github.com/yakworks/gorm-tools/issues/86)
+ - ParallelTools and GparsParallelTools need tests and docs [(#86)](https://github.com/yakworks/gorm-tools/issues/86)
  - 81 map binder [(#83)](https://github.com/yakworks/gorm-tools/pull/83)
  - GormMapBinder enhancments [(#82)](https://github.com/yakworks/gorm-tools/issues/82)
  - FastBinder refactor to a trait MapBinder [(#81)](https://github.com/yakworks/gorm-tools/issues/81)
- - Refactor to AsyncSupport so the gpars lib is behind the scenes a… [(#80)](https://github.com/yakworks/gorm-tools/pull/80)
+ - Refactor to ParallelTools so the gpars lib is behind the scenes a… [(#80)](https://github.com/yakworks/gorm-tools/pull/80)
  - Mango override for dao [(#79)](https://github.com/yakworks/gorm-tools/pull/79)
  - Code reformat [(#78)](https://github.com/yakworks/gorm-tools/pull/78)
  - DbDialectService working with H2 and tests [(#77)](https://github.com/yakworks/gorm-tools/issues/77)